Move compat tests to colocated files in src/providers/
Split test/compat.test.js into src/providers/openai.compat.test.js and
src/providers/anthropic.compat.test.js to follow the project convention
of colocating tests next to their source files.
