+++ src/providers/openai.compat.test.js
@@ -1,22 +1,18 @@
-// SDK compatibility tests.
-// These run in CI against multiple versions of the OpenAI and Anthropic SDKs.
-// They inject a mock fetch via the SDK's constructor option, let the real SDK
-// parse responses into its own objects, then verify our extract functions
-// handle those objects correctly.
-// Skipped locally if SDKs aren't installed.
+// SDK compatibility tests for OpenAI.
+// These use the real OpenAI SDK with a mock fetch to verify our extract
+// functions handle real SDK response objects (with getters, _client refs, etc).
+// Skipped if the OpenAI SDK isn't installed.
 
 import { describe, it, expect, vi } from 'vitest';
-import { warp, ref, flush } from '../src/index.js';
-import { responseRegistry } from '../src/core/registry.js';
-import { setupBeforeEach } from './setup.js';
+import { warp, ref, flush } from '../index.js';
+import { responseRegistry } from '../core/registry.js';
+import { setupBeforeEach } from '../../test/setup.js';
 
-let OpenAI, Anthropic;
-let openaiProvider, anthropicProvider;
+let OpenAI;
+let openaiProvider;
 
 try { ({ default: OpenAI } = await import('openai')); } catch {}
-try { ({ default: Anthropic } = await import('@anthropic-ai/sdk')); } catch {}
-try { openaiProvider = await import('../src/providers/openai.js'); } catch {}
-try { anthropicProvider = await import('../src/providers/anthropic.js'); } catch {}
+try { openaiProvider = await import('./openai.js'); } catch {}
 
 setupBeforeEach();
 
@@ -52,16 +48,6 @@ const OPENAI_RESPONSES_PAYLOAD = {
   usage: { input_tokens: 8, output_tokens: 6, total_tokens: 14 },
 };
 
-const ANTHROPIC_PAYLOAD = {
-  id: 'msg-compat-test',
-  type: 'message',
-  role: 'assistant',
-  model: 'claude-3-5-sonnet-latest',
-  content: [{ type: 'text', text: 'Hello from Anthropic compat!' }],
-  stop_reason: 'end_turn',
-  usage: { input_tokens: 12, output_tokens: 8 },
-};
-
 // ---------------------------------------------------------------------------
 // Helper: create a mock fetch that returns a JSON payload
 // ---------------------------------------------------------------------------
@@ -88,7 +74,7 @@ const hasResponsesAPI = OpenAI
   : false;
 
 // ---------------------------------------------------------------------------
-// OpenAI
+// Tests
 // ---------------------------------------------------------------------------
 
 describe.skipIf(!OpenAI)('OpenAI SDK compatibility', () => {
@@ -152,56 +138,6 @@ describe.skipIf(!OpenAI)('OpenAI SDK compatibility', () => {
 
     expect(mockFetch).toHaveBeenCalled();
     expect(result.choices[0].message.content).toBe('Hello from compat test!');
-    expect(responseRegistry.has(result)).toBe(true);
-    expect(ref(result)).toMatch(/^wm_call_/);
-  });
-});
-
-// ---------------------------------------------------------------------------
-// Anthropic
-// ---------------------------------------------------------------------------
-
-describe.skipIf(!Anthropic)('Anthropic SDK compatibility', () => {
-  it('detects Anthropic via constructor name', () => {
-    const client = new Anthropic({ apiKey: 'sk-ant-test' });
-    expect(client.constructor.name).toBe('Anthropic');
-  });
-
-  it('client has messages.create', () => {
-    const client = new Anthropic({ apiKey: 'sk-ant-test' });
-    expect(typeof client.messages.create).toBe('function');
-  });
-
-  it('extract handles real SDK response', async () => {
-    const mockFetch = createMockFetch(ANTHROPIC_PAYLOAD);
-    const client = new Anthropic({ apiKey: 'sk-ant-test', fetch: mockFetch });
-
-    const result = await client.messages.create({
-      model: 'claude-3-5-sonnet-latest',
-      max_tokens: 100,
-      messages: [{ role: 'user', content: 'hi' }],
-    });
-
-    expect(mockFetch).toHaveBeenCalled();
-    const ext = anthropicProvider.extract(result);
-    expect(ext.response).toBe('Hello from Anthropic compat!');
-    expect(ext.tokens.prompt).toBe(12);
-    expect(ext.tokens.completion).toBe(8);
-  });
-
-  it('warp intercepts and tracks a real SDK call', async () => {
-    const mockFetch = createMockFetch(ANTHROPIC_PAYLOAD);
-    const client = new Anthropic({ apiKey: 'sk-ant-test', fetch: mockFetch });
-    const wrapped = warp(client);
-
-    const result = await wrapped.messages.create({
-      model: 'claude-3-5-sonnet-latest',
-      max_tokens: 100,
-      messages: [{ role: 'user', content: 'hi' }],
-    });
-
-    expect(mockFetch).toHaveBeenCalled();
-    expect(result.content[0].text).toBe('Hello from Anthropic compat!');
     expect(responseRegistry.has(result)).toBe(true);
     expect(ref(result)).toMatch(/^wm_call_/);
   });
